added theme panel and configuration settings
--the user can choose that instead of the default barLine,
he would like to use the theme's one. This is something very
important in cases where the theme supports transparency
for a specific panel (e.g. west, east, north, south) and
not all of them... This technique can be used in order to
give the user an illusion that the icons grow and zoom
outside of their panel